refactor: tic-tac-toeのリファクタ途中まで
#44
