+++ testing_basic/tic-tac-toe/src/Square/Square.js
@@ -1,0 +1,7 @@
+export const Square = (props) => {
+  return (
+    <button className="square" onClick={props.onClick}>
+      {props.value}
+    </button>
+  );
+};
