+++ testing_basic/tic-tac-toe/src/stories/game.stories.js
@@ -1,5 +1,5 @@
 import { React } from "react";
-import { Game } from "../index";
+import { Game } from "../Game/Game";
 
 export default {
   title: "Tic-Tac-Toe/Game",
